+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [4.0.0] -
 
+### Chnaged
+
+- Replace legacy endpoints
+
 ## [3.3.1] - 2022-07-13
 
 ### Changed