+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.1] - 2022-0
+
+### Changed
+
 ## [4.0.0] - 2022-07-29
 
 ### Added
@@ -429,6 +433,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add RFC cross chain navigation
 - Add a switch to selcet a dark theme from navigation drawer
 
+[4.0.1]: https://github.com/commercionetwork/almerico/compare/v4.0.0...v4.0.1
 [4.0.0]: https://github.com/commercionetwork/almerico/compare/v3.3.1...v4.0.0
 [3.3.1]: https://github.com/commercionetwork/almerico/compare/v3.3.0...v3.3.1
 [3.3.0]: https://github.com/commercionetwork/almerico/compare/v3.1.6...v3.3.0